+++ packages/react-components-lab/src/components/Card/types.ts
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react';
+
+export interface CardProps {
+  title: string;
+  description?: string[] | string | ReactNode;
+  isOpen?: boolean;
+  subTitle?: string | null;
+  setIsOpen?: (d: boolean) => void;
+  image?: string;
+  children?: ReactNode;
+  isClickable?: boolean;
+  disabled?: boolean;
+}