+++ packages/react-components-lab/src/components/GapBox/GapBox.stories.tsx
@@ -13,7 +13,7 @@ export default {
 const Template: Story<Props> = (args) => <GapBox {...args} />;
 
 const Chips = ['Ocean', 'Waves', 'Sea', 'Tides'].map((label) => (
-  <Chip label={label} />
+  <Chip key={`chip-${label}`} label={label} />
 ));
 
 export const Flex = Template.bind({});