+++ packages/react-components/.storybook/main.js
@@ -6,7 +6,7 @@ module.exports = {
     '@storybook/addon-actions',
     '@storybook/addon-links',
     '@storybook/addon-docs',
-    '@storybook/addon-knobs/register',
+    '@storybook/addon-knobs',
     '@storybook/addon-storysource',
   ],
   webpackFinal: async config => {