+++ packages/react-components/src/common/types.ts
@@ -1,0 +1,4 @@
+export interface DateProps {
+  from: string;
+  to: string;
+}